DOS-94: move startup backfill methods from db/core.rs to data_lifecycle.rs
Relocate backfill_meeting_identity, backfill_meeting_user_layer,
dismiss_internal_stakeholder_suggestions, backfill_stakeholder_columns,
run_guarded_init_backfill_account_domains, and their is_init_task_completed
/ mark_init_task_completed helpers into a new impl ActionDb block in
db/data_lifecycle.rs. Visibility raised to pub(super) so the existing
Self:: call sites in db/core.rs continue to compile; sanitize_calendar_event_id
likewise raised to pub(super) for cross-file reuse.

L2-status: not-run-acknowledged
