+++ src-tauri/src/db/core.rs
@@ -513,398 +513,8 @@ impl ActionDb {
             .to_string()
     }
 
-    fn sanitize_calendar_event_id(calendar_event_id: &str) -> String {
+    pub(super) fn sanitize_calendar_event_id(calendar_event_id: &str) -> String {
         calendar_event_id.replace('@', "_at_")
-    }
-
-    /// Re-key meeting IDs to canonical event IDs and update dependent references.
-    fn backfill_meeting_identity(conn: &Connection) -> Result<(), DbError> {
-        let rows: Vec<(String, String)> = {
-            let mut stmt = conn.prepare(
-                "SELECT id, calendar_event_id
-                 FROM meetings
-                 WHERE calendar_event_id IS NOT NULL
-                   AND trim(calendar_event_id) != ''",
-            )?;
-            let mapped = stmt.query_map([], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-            })?;
-            let mut items = Vec::new();
-            for row in mapped {
-                items.push(row?);
-            }
-            items
-        };
-
-        for (old_id, calendar_event_id) in rows {
-            let canonical_id = Self::sanitize_calendar_event_id(&calendar_event_id);
-            if canonical_id.is_empty() || canonical_id == old_id {
-                continue;
-            }
-
-            let canonical_exists: i64 = conn.query_row(
-                "SELECT COUNT(*) FROM meetings WHERE id = ?1",
-                params![canonical_id],
-                |r| r.get(0),
-            )?;
-
-            if canonical_exists > 0 {
-                // Merge sparse fields from old row into canonical row (meetings table).
-                conn.execute(
-                    "UPDATE meetings
-                     SET title = COALESCE(title, (SELECT title FROM meetings WHERE id = ?1)),
-                         meeting_type = COALESCE(meeting_type, (SELECT meeting_type FROM meetings WHERE id = ?1)),
-                         start_time = COALESCE(start_time, (SELECT start_time FROM meetings WHERE id = ?1)),
-                         end_time = COALESCE(end_time, (SELECT end_time FROM meetings WHERE id = ?1)),
-                         attendees = COALESCE(attendees, (SELECT attendees FROM meetings WHERE id = ?1)),
-                         notes_path = COALESCE(notes_path, (SELECT notes_path FROM meetings WHERE id = ?1)),
-                         description = COALESCE(description, (SELECT description FROM meetings WHERE id = ?1))
-                     WHERE id = ?2",
-                    params![old_id, canonical_id],
-                )?;
-                // Merge meeting_prep fields
-                conn.execute(
-                    "UPDATE meeting_prep
-                     SET prep_context_json = COALESCE(prep_context_json, (SELECT prep_context_json FROM meeting_prep WHERE meeting_id = ?1)),
-                         user_agenda_json = COALESCE(user_agenda_json, (SELECT user_agenda_json FROM meeting_prep WHERE meeting_id = ?1)),
-                         user_notes = COALESCE(user_notes, (SELECT user_notes FROM meeting_prep WHERE meeting_id = ?1)),
-                         prep_frozen_json = COALESCE(prep_frozen_json, (SELECT prep_frozen_json FROM meeting_prep WHERE meeting_id = ?1)),
-                         prep_frozen_at = COALESCE(prep_frozen_at, (SELECT prep_frozen_at FROM meeting_prep WHERE meeting_id = ?1)),
-                         prep_snapshot_path = COALESCE(prep_snapshot_path, (SELECT prep_snapshot_path FROM meeting_prep WHERE meeting_id = ?1)),
-                         prep_snapshot_hash = COALESCE(prep_snapshot_hash, (SELECT prep_snapshot_hash FROM meeting_prep WHERE meeting_id = ?1))
-                     WHERE meeting_id = ?2",
-                    params![old_id, canonical_id],
-                )?;
-                // Merge meeting_transcripts fields
-                conn.execute(
-                    "UPDATE meeting_transcripts
-                     SET summary = COALESCE(summary, (SELECT summary FROM meeting_transcripts WHERE meeting_id = ?1)),
-                         transcript_path = COALESCE(transcript_path, (SELECT transcript_path FROM meeting_transcripts WHERE meeting_id = ?1)),
-                         transcript_processed_at = COALESCE(transcript_processed_at, (SELECT transcript_processed_at FROM meeting_transcripts WHERE meeting_id = ?1))
-                     WHERE meeting_id = ?2",
-                    params![old_id, canonical_id],
-                )?;
-            } else {
-                conn.execute(
-                    "UPDATE meetings SET id = ?1 WHERE id = ?2",
-                    params![canonical_id, old_id],
-                )?;
-                // Child tables updated via CASCADE on meetings.id
-            }
-
-            // Update foreign references.
-            conn.execute(
-                "UPDATE captures SET meeting_id = ?1 WHERE meeting_id = ?2",
-                params![canonical_id, old_id],
-            )?;
-            conn.execute(
-                "UPDATE meeting_entities SET meeting_id = ?1 WHERE meeting_id = ?2",
-                params![canonical_id, old_id],
-            )?;
-            conn.execute(
-                "UPDATE meeting_attendees SET meeting_id = ?1 WHERE meeting_id = ?2",
-                params![canonical_id, old_id],
-            )?;
-            conn.execute(
-                "UPDATE actions
-                 SET source_id = ?1
-                 WHERE source_type IN ('transcript', 'post_meeting') AND source_id = ?2",
-                params![canonical_id, old_id],
-            )?;
-
-            // Update reviewed state keys.
-            conn.execute(
-                "UPDATE meeting_prep_state
-                 SET prep_file = ?1
-                 WHERE prep_file = ?2 OR prep_file = ?3",
-                params![canonical_id, old_id, format!("preps/{}.json", old_id)],
-            )?;
-
-            if canonical_exists > 0 {
-                // CASCADE will clean up meeting_prep and meeting_transcripts
-                conn.execute("DELETE FROM meetings WHERE id = ?1", params![old_id])?;
-            }
-        }
-        Ok(())
-    }
-
-    fn backfill_meeting_user_layer(conn: &Connection) -> Result<(), DbError> {
-        let rows: Vec<(String, String, Option<String>, Option<String>)> = {
-            let mut stmt = conn.prepare(
-                "SELECT mp.meeting_id, mp.prep_context_json, mp.user_agenda_json, mp.user_notes
-                 FROM meeting_prep mp
-                 WHERE mp.prep_context_json IS NOT NULL
-                   AND trim(mp.prep_context_json) != ''",
-            )?;
-            let mapped = stmt.query_map([], |row| {
-                Ok((
-                    row.get::<_, String>(0)?,
-                    row.get::<_, String>(1)?,
-                    row.get::<_, Option<String>>(2)?,
-                    row.get::<_, Option<String>>(3)?,
-                ))
-            })?;
-            let mut items = Vec::new();
-            for row in mapped {
-                items.push(row?);
-            }
-            items
-        };
-
-        for (meeting_id, prep_json, agenda_existing, notes_existing) in rows {
-            let Ok(value) = serde_json::from_str::<serde_json::Value>(&prep_json) else {
-                continue;
-            };
-            let agenda_from_prep = value
-                .get("userAgenda")
-                .and_then(|v| v.as_array())
-                .map(|arr| {
-                    arr.iter()
-                        .filter_map(|v| v.as_str().map(|s| s.trim().to_string()))
-                        .filter(|s| !s.is_empty())
-                        .collect::<Vec<String>>()
-                })
-                .filter(|v| !v.is_empty())
-                .and_then(|v| serde_json::to_string(&v).ok());
-            let notes_from_prep = value
-                .get("userNotes")
-                .and_then(|v| v.as_str())
-                .map(|s| s.trim().to_string())
-                .filter(|s| !s.is_empty());
-
-            let agenda_target = agenda_existing.or(agenda_from_prep);
-            let notes_target = notes_existing.or(notes_from_prep);
-            if agenda_target.is_none() && notes_target.is_none() {
-                continue;
-            }
-
-            conn.execute(
-                "UPDATE meeting_prep
-                 SET user_agenda_json = COALESCE(user_agenda_json, ?1),
-                     user_notes = COALESCE(user_notes, ?2)
-                 WHERE meeting_id = ?3",
-                params![agenda_target, notes_target, meeting_id],
-            )?;
-        }
-        Ok(())
-    }
-
-    /// Auto-dismiss stakeholder suggestions for internal team members.
-    /// Cleans up suggestions that were created before the internal filter was added.
-    fn dismiss_internal_stakeholder_suggestions(conn: &Connection) -> Result<(), DbError> {
-        let dismissed = conn.execute(
-            "UPDATE stakeholder_suggestions SET status = 'dismissed', resolved_at = datetime('now')
-             WHERE status = 'pending' AND (
-               (person_id IS NOT NULL AND EXISTS (
-                 SELECT 1 FROM people p WHERE p.id = stakeholder_suggestions.person_id AND p.relationship = 'internal'
-               ))
-               OR (suggested_email IS NOT NULL AND EXISTS (
-                 SELECT 1 FROM people p2 WHERE p2.email = LOWER(stakeholder_suggestions.suggested_email) AND p2.relationship = 'internal'
-               ))
-               OR (suggested_name IS NOT NULL AND EXISTS (
-                 SELECT 1 FROM people p3 WHERE LOWER(p3.name) = LOWER(stakeholder_suggestions.suggested_name) AND p3.relationship = 'internal'
-               ))
-             )",
-            [],
-        )?;
-        if dismissed > 0 {
-            log::info!(
-                "I652: auto-dismissed {} internal stakeholder suggestions",
-                dismissed
-            );
-        }
-        Ok(())
-    }
-
-    /// Backfill engagement/assessment columns on `account_stakeholders` from
-    /// the legacy `entity_assessment.stakeholder_insights_json` blob.
-    /// Runs once at startup — only touches rows where `engagement IS NULL`.
-    fn backfill_stakeholder_columns(conn: &Connection) -> Result<(), DbError> {
-        // Step 1: Find all account_stakeholders rows missing engagement.
-        let rows: Vec<(String, String)> = {
-            let mut stmt = conn.prepare(
-                "SELECT account_id, person_id FROM account_stakeholders WHERE engagement IS NULL",
-            )?;
-            let mapped = stmt.query_map([], |row| {
-                Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-            })?;
-            let mut items = Vec::new();
-            for row in mapped {
-                items.push(row?);
-            }
-            items
-        };
-
-        if rows.is_empty() {
-            return Ok(());
-        }
-
-        // Step 2: Group by account_id to avoid repeated JSON parses.
-        let mut by_account: std::collections::HashMap<String, Vec<String>> =
-            std::collections::HashMap::new();
-        for (account_id, person_id) in &rows {
-            by_account
-                .entry(account_id.clone())
-                .or_default()
-                .push(person_id.clone());
-        }
-
-        let mut updated = 0u32;
-        for (account_id, person_ids) in &by_account {
-            // Step 3: Read the stakeholder_insights_json for this entity.
-            let json_opt: Option<String> = conn
-                .query_row(
-                    "SELECT stakeholder_insights_json FROM entity_assessment WHERE entity_id = ?1",
-                    params![account_id],
-                    |row| row.get(0),
-                )
-                .ok();
-
-            let json_str = match json_opt {
-                Some(ref s) if !s.is_empty() => s.as_str(),
-                _ => continue,
-            };
-
-            let entries: Vec<serde_json::Value> = match serde_json::from_str(json_str) {
-                Ok(v) => v,
-                Err(err) => {
-                    log::warn!(
-                        "I652 backfill: failed to parse stakeholder_insights_json for {}: {}",
-                        account_id,
-                        err
-                    );
-                    continue;
-                }
-            };
-
-            // Step 4: For each person_id, find a matching entry and update.
-            for person_id in person_ids {
-                let matching = entries.iter().find(|e| {
-                    e.get("person_id")
-                        .and_then(|v| v.as_str())
-                        .map(|pid| pid == person_id)
-                        .unwrap_or(false)
-                });
-
-                if let Some(entry) = matching {
-                    let engagement = entry.get("engagement").and_then(|v| v.as_str());
-                    let assessment = entry.get("assessment").and_then(|v| v.as_str());
-
-                    if engagement.is_some() || assessment.is_some() {
-                        conn.execute(
-                            "UPDATE account_stakeholders
-                             SET engagement = COALESCE(engagement, ?1),
-                                 assessment = COALESCE(assessment, ?2),
-                                 data_source_engagement = COALESCE(data_source_engagement, 'ai'),
-                                 data_source_assessment = COALESCE(data_source_assessment, 'ai')
-                             WHERE account_id = ?3 AND person_id = ?4
-                               AND engagement IS NULL",
-                            params![engagement, assessment, account_id, person_id],
-                        )?;
-                        updated += 1;
-                    }
-                }
-            }
-        }
-
-        if updated > 0 {
-            log::info!(
-                "I652 backfill: populated engagement/assessment for {} stakeholder rows",
-                updated
-            );
-        }
-        Ok(())
-    }
-
-    /// Check if a one-time init task has been completed.
-    ///
-    /// Returns true if the task has already run and been marked in init_tasks.
-    fn is_init_task_completed(conn: &Connection, task_name: &str) -> Result<bool, DbError> {
-        let completed = conn
-            .query_row(
-                "SELECT 1 FROM init_tasks WHERE task_name = ?1",
-                params![task_name],
-                |_| Ok(true),
-            )
-            .unwrap_or(false);
-        Ok(completed)
-    }
-
-    /// Mark a one-time init task as completed.
-    fn mark_init_task_completed(conn: &Connection, task_name: &str) -> Result<(), DbError> {
-        conn.execute(
-            "INSERT OR IGNORE INTO init_tasks (task_name) VALUES (?1)",
-            params![task_name],
-        )?;
-        Ok(())
-    }
-
-    /// Guarded backfill: Account domains from meeting attendees (Path 2b entity resolution).
-    ///
-    /// Runs exactly once. Subsequent calls are guarded by init_tasks table.
-    fn run_guarded_init_backfill_account_domains(&self) -> Result<(), DbError> {
-        // v3: purge ALL domains (v1 and v2 both stored contaminated data),
-        // then re-backfill with domain-base matching (only stores a domain
-        // on an account when the domain base matches the account name/slug).
-        const TASK_NAME: &str = "backfill_account_domains_v3";
-
-        if Self::is_init_task_completed(&self.conn, TASK_NAME)? {
-            return Ok(());
-        }
-
-        // Purge all contaminated domains — the new backfill will repopulate correctly
-        let purged = self.conn.execute("DELETE FROM account_domains", [])?;
-        if purged > 0 {
-            log::info!(
-                "Entity resolution: purged all {} account_domains rows (v3 clean slate)",
-                purged,
-            );
-        }
-
-        // Also clean up over-linked future meetings (>2 entity links = contamination symptom)
-        let cleaned = self.conn.execute(
-            "DELETE FROM meeting_entities WHERE meeting_id IN (
-                SELECT me.meeting_id FROM meeting_entities me
-                JOIN meetings m ON m.id = me.meeting_id
-                WHERE m.start_time > datetime('now')
-                GROUP BY me.meeting_id HAVING COUNT(*) > 2
-            )",
-            [],
-        )?;
-        if cleaned > 0 {
-            log::info!(
-                "Entity resolution: purged {} over-linked future meeting_entities rows",
-                cleaned,
-            );
-            // Clear prep for those meetings so they regenerate
-            #[allow(
-                clippy::let_underscore_must_use,
-                reason = "intentional best-effort discard; preserves existing non-blocking behavior"
-            )]
-            // best-effort: orphan cleanup succeeded; stale prep is regenerated by later prep invalidation paths.
-            let _ = self.conn.execute(
-                "UPDATE meeting_prep SET prep_frozen_json = NULL, prep_context_json = NULL
-                 WHERE meeting_id IN (
-                     SELECT id FROM meetings WHERE start_time > datetime('now')
-                     AND id NOT IN (SELECT meeting_id FROM meeting_entities)
-                 )",
-                [],
-            );
-        }
-
-        // Re-backfill with domain-base matching
-        let inserted = self.backfill_account_domains_from_meetings()?;
-
-        Self::mark_init_task_completed(&self.conn, TASK_NAME)?;
-
-        if inserted > 0 {
-            log::info!(
-                "Entity resolution: backfilled {} account→domain mappings (domain-base matching)",
-                inserted
-            );
-        }
-
-        Ok(())
     }
 
     /// One-time backfill of dashboard.json narrative fields into DB columns.
